requirements.txt
imageio==2.9.0
matplotlib==3.5.0
numpy==1.20.1
Pillow==9.2.0
POT==0.8.1.0
torch==1.8.0
torchvision==0.9.0
